Fixes bug that chopped large buffers of text
If a text element contains enough text such that it requires multiple
stream writes (i.e. exceeds the capacity of the ReadableStream buffer),
then the text previously read from the buffer would be lost. This patch
solves that issue by appending to the buffer to the node's text property.

Additionally, it fixes an issue where whitespace that was part of the
formatting within text was lost. It does this by altering the location where
string trims are performed. String trims are now done within the
"end element" callback.